Added messaging between Order, Product and Payment services
Now Payment service also caches the Order inside a redis database. If the order is validated by the ProductService, then it can be sent. If the order stock can't be reserved, the order gets automatically cancelled and the user cannot continue to the payment.
ProductService now sends messages to both Order and Payment service for confirming the order.
Documented the code.